- Решил задачу, предложенную на уроке Вот условие:
Binary search
Implement a binary search algorithm.

Input
The first line of the input contains the natural numbers N and K (0<N,K<20,000).
The second line specifies the N elements of the first array sorted in ascending
order, and the third line specifies the K elements of the second array.
The elements of both arrays are integers, each of which does not exceed 109
in absolute value

Output
It is required to output YES for each of K numbers in a separate line,
if this number occurs in the first array, and NO otherwise.

        Scanner scanner = new Scanner(System.in);
        String[] str = scanner.nextLine().split(" ");
        int n = Integer.parseInt(str[0]);
        int k = Integer.parseInt(str[1]);